cm: reorganize tests
